re-organize components based on function
