+++ src/App.scss
@@ -13,3 +13,4 @@ $input-focus-box-shadow: $input-box-shadow; // hack to get upgrade to paragon 4.
 
 @import "./components/GradesTab/GradesTab";
 @import "./components/WithSidebar/WithSidebar";
+@import "./components/GradebookFilters/GradebookFilters";
