+++ src/components/EdxHeader/index.jsx
@@ -3,11 +3,11 @@ import { Hyperlink } from '@edx/paragon';
 import { getConfig } from '@edx/frontend-platform';
 
 /**
- * <Header />
+ * <EdxHeader />
  * Gradebook MFE app header.
  * Displays edx logo, linked to lms dashboard
  */
-const Header = () => (
+const EdxHeader = () => (
   <div className="mb-3">
     <header className="d-flex justify-content-center align-items-center p-3 border-bottom-blue">
       <Hyperlink destination={`${getConfig().LMS_BASE_URL}/dashboard`}>
@@ -18,4 +18,4 @@ const Header = () => (
   </div>
 );
 
-export default Header;
+export default EdxHeader;
