+++ src/components/GradebookFilters/GradebookFilters.scss
@@ -1,0 +1,6 @@
+.filter-sidebar-header {
+    display: flex;
+    align-items: flex-start;
+    justify-content: space-between;
+    padding: 15px;
+}
