+++ src/components/WithSidebar/WithSidebar.scss
@@ -12,18 +12,6 @@ $sidebar-width: 350px;
     }
 }
 
-.sidebar-contents {
-    overflow-x: auto;
-    transition: margin 300ms cubic-bezier(0.4,0,0.2,1);
-    margin-left: 0;
-    .sidebar.open + & {
-        margin-left: $sidebar-width;
-    }
-    &.opening {
-        width: calc(100vw - #{$sidebar-width});
-    }
-}
-
 .sidebar-header {
     display: flex;
     align-items: flex-start;
